Add capability checks to non-REST write operations, harden input
Security audit findings:

- PatchBlockAbility: add edit_post check before wp_update_post
- DuplicatePostAbility: add edit_post + create_posts checks
- ManageRevisionsAbility: add edit_post check before restore
- CreateTranslationAbility: add edit_post + create_posts checks
- CreateTermTranslationAbility: add edit_terms check
- UpdateStringTranslationAbility: require manage_options
- ManageRedirectsAbility: require publish_pages for create action
- BulkUpdatePostsAbility: block protected meta keys via is_protected_meta()
- Cast all user-provided IDs to (int) across all abilities

REST-delegating abilities (content CRUD, terms) already enforce
capabilities via rest_do_request() — the REST controller's own
permission_callback runs. These are not affected.

12 security tests verify that subscribers cannot:
- patch blocks, duplicate posts, restore revisions
- create translations, create redirects
- write protected meta keys

+++ src/Integrations/Polylang/MachineTranslateAbility.php
@@ -108,7 +108,7 @@ final class MachineTranslateAbility
             return self::translateStrings($input['string_group'], $targetLang, $service, $language);
         }
 
-        $postId = $input['id'] ?? 0;
+        $postId = (int) ($input['id'] ?? 0);
         if (! $postId) {
             return new WP_Error('missing_input', 'Provide post_id or string_group. For terms, use gds/translations-create-term instead.');
         }
